Several improvements in the offloading support.
The driver now forwards the host IR to the the target compiler phase so
that it can retrieve all the target entries as well as any other
declaration that has to be emitted for the target, in the same order
they are created for the host. This information is recorded in the form
of metadata during the host code generation. This allowed us to highly
simplify the checks that were being done in the target side to
understand what declarations should be emitted and tackle several issues
reported by our current users, mainly in the emission of deferred and
template  declarations.

The codegen scheme for the GPU was improved to get better performance,
support -O3 optimization, and make it less prone to incompatibilities
with the NVPTX toolchain, namely the PTX assembler.

Use a less memory hungry scheme to share data between GPU threads.

Fix in the visibility of offloading related symbols generated by the
compiler and linker (offloading linker script update) in order to
correctly register target regions from shared libraries.

Enable LTO for offloading code. The LTO commands are created by the
driver given that the gold linker does not understand the linker script
required to drive the host-target image linkage.